Adding a button to download reports
We are implementing a new feature where a BEIS user can download all relevant
reports with only clicking on a button. This will save time, as they don't need
to go to each report page and download it individually.

This commit adds the button that will trigger the CSV download on the reports
index page.