Set a transaction's financial quarter from its date
We're going to migrate transactions from having an exact date, to having
a financial quarter. To do this we need to set the financial quarter for
existing transactions, based on their date. Once the system is assigning
financial quarters for all created/updated transactions, we can
back-fill this data for all the others.