Back-fill financial quarter for transactions
This migration fills in the inferred financial quarter for all
transactions based on their date. When this is deployed, no transactions
will have a financial quarter set, so all transactions with dates should
be updated in this way.

At time of writing, all the transactions in production have a date set,
so all of them will be updated by this migration.

We need to use `update_columns` here because not all transactions in
production pass validation, so they'd fail to save if we used `update!`.