Change interface of new upload page for actuals
Move the CSV download link into the main page content and add some guidance.

This is the simplest change we could mae to the design to validate a more directed journey for uploading bulk transaction data: the main elements are making the download link more prominent, and giving the same weight to the download and upload elements.