Create programmes through parent fund rather than source_fund
After some consideration, I decided it would be better if new child
activities were created through a consistent URL, rather than treating
funds differently.

This should also prevent some very confusing problems with testing,
because it was very easy to call `authenticate!` before the fund
activities were initialised. Now we're looping over the fund activities,
rather than `Fund.all` (which always exists, as it is loaded from a
codelist file), we won't render the 'Add fund to...' links during the
sign-in phase, causing a failure, but they will exist when we actually
want to test them!

If we choose to treat funds as a distinct entity in the future, it
should be straightforward enough to split this out into its own
controller.