Users can add a child activity for an existing programme
Reuse ActivityChildrenController#create to create projects from an
organisation's programme list.

This also reuses `policy(:project).create?`, which revealed some
inconsistencies between the main `ActivityPolicy` and the level-specific
policies (which are used interchangeably). It also doesn't currently check
for an active report, so they'll be times cases where the link is
displayed, but the user receives an `HTTP 500` when clicking on it, but
fixing the policies was a rabbit hole I wanted to avoid here.