Handle the BOM on uploaded files
This adds a new `CsvFileUpload` class which params with an uploaded file, reads
the file and optionally removes the Byte Order Mark, which is added by Excel.
There are shortcuts to doing this in Ruby, but we felt that it was more sensible to
document it as a problem we're aware of in code.

I've also updated some of the tests, which expect files with invalid encoding to be
successfully read by the CSV parser, which is not the case.