Set defaults on budgets
This forms one of the many changes to extend budgets to be useful as the
main way to describe many funding arrangements in the application.

The IATI fields budget type and budget status are for the time being
going to always be 'original' and 'committed', so we have no need to ask
for them and can set their values on new budgets.

We don't need to worry about any budget we already have that might have
other status or types as there is other work that will happen to resolve
those issues planned, we think all budgets will be deleted as they don't
reflect the new concept of budget that allows BEIS to collect funding
arrangements anyway.

By keeping the fields we an always change our minds and locate any
budget that do not meet this new criteria.