Switch transaction validation from dates to quarters
To migrate the transaction input UI over from dates to financial
quarters, the validation needs to be in terms of quarters -- it's no use
giving the user an error about a date when they're not entering a date.

So here we make the following changes to transaction validation:

- require `financial_year` to be present
- require `financial_quarter` to be a value from 1 to 4
- do not require `date` to be present
- do not require `date` to be in the past

The last point is because of how we generate dates: the date is set to
the final day of the transaction's quarter. Given that transactions can
be entered during the quarter in which they took place, that means they
may be assigned a date in the future.

We also change how `set_financial_quarter_from_date` works so that it
can work in both directions: if the transaction has a date it's assigned
a quarter, and vice versa. This helps with keeping all methods of
transaction input working through this transition.