Migrate the transaction web form to use financial quarters
In the last few commits we migrated the transaction model to support
storing a financial quarter instead of a date. This commit migrates the
web forms for entering a single transaction over to that way of working.

The new transaction's quarter is defaulted to the quarter of the current
editable report for the activity, if one exists.