Migrate the transaction bulk importer to use financial quarters
In the last few commits we migrated the transaction model to support
storing a financial quarter instead of a date. This commit migrates the
bulk importer for entering multiple transactions over to that way of
working.

Each transaction's quarter in the downloadable CSV template is defaulted
to the quarter of the current editable report for the activity, if one
exists.