Set default funding type
When a new budget is created the budget type should default to that of
the parent activity's source fund i.e. a Newton Activity should have
Newton budget and a GCRF only GCRF budget.

Budgets do not yet enforce this (although that will happen soon) so we
set the default value for the moment.