Validate funding_type matches source_fund
This is a stepping stone to a complete concept for budget in the
application.

When adding budgets to an activity the funding_type should only ever be
the same as the parent activity's source_fund i.e. Newton fund
activities should only have Newton budget.

Future work will introduce funding_types to budget beyond those we have
now, therefore we want to keep the UI that asks the user for the funding
type but we do not want to allow users to add the wrong budget fund
types to activities in the mean time.

The validation can only be performed if the parent activity has a
`source_fund_code` so we check for the presence of that attribute. Work
happening elsewhere will ensure that activities cannot be created
without this attribute and in the meantime the budget policy prevent a
budget being added when the parent activity is in the incorrect state.

As we any change to validation, some specs needed to be update to ensure
budgets are being created in the appropriate state for the test.