Validate funding_type matches source_fund
This is a stepping stone to a complete concept for budget in the
application.

When adding budgets to an activity the funding_type should only ever be
the same as the parent activity's source_fund i.e. Newton fund
activities should only have Newton budget.

Future work will introduce funding_types to budget beyond those we have
now, therefore we want to keep the UI that asks the user for the funding
type but we do not want to allow users to add the wrong budget fund
types to activities in the mean time.

As we any change to validation, some specs needed to be update to ensure
budgets are being created in the appropriate state for the test.