Introduce HasFinancialQuarter
We have a few places in the codebase that repeat the logic for
constructing a FinancialQuarter object from a record's
`financial_quarter` and `financial_year` fields, and generating the
string representation of these.

This module replaces all these instances with a common implementation.
To avoid confusion with the integer `financial_quarter` attribute, the
method for getting a FinancialQuarter is named `own_financial_quarter`.
This name choice is intended to contrast with other uses we anticipate
such as `next_financial_quarter`.

All uses of the method `financial_quarter_and_year` now return a string
beginning with `FQ`, whereas some returned just `Q` previously, for
consistency with `Report#financial_quarter_and_year`.