Convert FinancialQuarter objects back into query parameters
By adding a `to_hash` method to FinancialQuarter, we can use it to
generate keyword arguments using the `**` operator and thus pass a
FinancialQuarter as input to ActiveRecord queries. This tidies up a fair
bit of repetitive logic for threading these values through code.

This also changes the implementation of
`Activity#actual_total_for_report_financial_quarter`. Rather than using
the `created_at` field on Report, we use its `financial_{quarter,year}`
fields. These are set to the current quarter when the Report is created
and should be the canonical source to determine what quarter a Report
relates to.

We can also combine this with the `FinancialQuarter#pred` method to tidy
up the implementation of `Report#previous`.