Make PlannedDisbursementHistory take the financial quarter as keyword arguments
Instead of taking the quarter and year as positional parameters, we make
PlannedDisbursementHistory take keyword arguments. This makes calls to
`PlannedDisbursementHistory.new` clearer and makes its interface
consistent with `PlannedDisbursementOverview::Snapshot#value_for`.

It also means that FinancialQuarter objects can more easily be used as
inputs to this class as they can be converted to keyword arguments.