Make FinancialQuarter return the following N quarters
We often want to generate the sequence of quarters following or
preceding a particular one. An instance of this is currently implemented
in `Report#next_twelve_financial_quarters` but it will be easier to make
some upcoming changes if this logic is extracted from the Report class
and put in FinancialQuarter instead.

This introduces `FinancialQuarter#following(n)` which returns the `n`
quarters that follow the receiver, using the `succ` method. We use this
to replace the aforementioned Report method, and tidy up some login in
ExportActivityToCsv where we can now pass FinancialQuarter objects into
`PlannedDisbursementOverview::Snapshot#value_for`.