Include previous 12 quarters of transactions in ExportActivityToCsv
Most of the work to make this possible is explained in the preceding
commits and this ends up being boilerplate. We replace the
`previous_quarter_actuals` logic with `previous_twelve_quarter_actuals`,
which uses TransactionOverview to load the transaction history that is
visible to the current report and query it.