Add a method to get all an Activity's descendants
Out of the box, `acts_as_tree` has a `descendants` method. However, this is slow,
as it manually walks down the tree and gets the child for each child.

This method overrides the `#descendants` method and instead uses Postgres's
`WITH RECURSIVE` functionality to walk down the tree and return all the
decendants in one query.