Remove test causing RSpec nil warning
A spec was checking for the presence of an activity's blank title (form step for this activity preceeded
one where a title is entered). This gave us the warning "Checking for expected text
of nil is confusing and/or pointless since it will always match." in the middle of
RSpec's output.