Financial quarter pretty name includes “F”
The requirements for filenames specify to display financial quarters in the format `FQ1 2020-2021`.

In order to maintain consistency across the application, all entities that have a financial quarter use the same method to display them.

This means that the most straightforward way to implement this change is to change “Q” to “FQ” everywhere, which includes filenames, tables, exported data, etc.

Previously: “Q4 2020-2021”

Now: “FQ4 2020-2021”