Previous report actuals header consistent
Use the report presenter to generate the header for the header for the previous report actuals.

A previous commit changed the display string for the financial quarter and year that is generated by the report presenter.

However, the export service wasn’t using the presenter for the financial quarter display of the previous quarter, and was instead generating it on the spot. This meant that any changes in the report presenter would have to be duplicated, when they are in fact governed by the same logic.