Expose BEIS organisation reference for editing
Now that the data has been pre-filled in production via the data migration, we can add validation for the presence and format of the organisation reference / short name.

Note: I have chosen to show the user input as it will be used, in uppercase letters, even if the user is typing lowercase letters. We are automatically converting the user input to uppercase before validating, to fulfill the requirements without making the user do the work that a machine can do.