Allow a Transfer to be created
This allows a BEIS user to create a Transfer record with a recieving Activity, a
financial period and a value by entering the RODA ID of the Activity and other
relevant details. There's a confirmation screen to make sure the activity is correct.
If the user clicks yes, the Transfer is created, if they click no, they're taken back to
the form with their previous options filled in.