Report presenter generates CSV filenames
The filenames for various types of downloadable CSVs have a lot in common: the financial quarter and year, the source fund’s short name, and the organisation’s short name.

Use the report presenter to generate all these filenames, with the individual variation for different purposes such as templates intended for the user to edit and upload, or downloads intended for reporting.