Users can add a child activity to an existing project
On the Activity "child activities" tab, change the "Add activity" button
so that it points to the new `ActivityChildren#create` action.

Because the new activity is automatically associated with its parent
activity and the `form_state` is set to beyond the 'level' and 'parent'
steps, the fill_in_activity_form test helper breaks.

Therefore, I've temporarily added a `skip_level_and_parent_steps` option
to the helper method, which will be removed once the old activity
creation code becomes obsolete.

To prevent breaking the audit-trail tests, I've added a
`create_activity` call in `ActivityChildrenController#create` which
mirrors the existing behaviour in `ActivitiesController#create`.