Optimise ExportActivityToCsv#next_twelve_quarter_forecasts
Currently, this method calls
`PlannedDisbursementOverview::Snapshot#value_for` once for each of the
twelve quarters it generates a forecast for. That's twelve queries for
every row that ExportActivityToCsv generates.

We can reduce this to one query by making
`PlannedDisbursementOverview::Snapshot#all_quarters` return an object
that loads all the forecasts for the activity (at their correct version
for the given report), and then provides a `value_for` method that uses
that in-memory result instead of sending a new query to the DB for a
single quarter's forecast.

We then create this object just once in `next_twelve_quarter_forecasts`
and call its `value_for` method for each quarter, so we do one query
instead of twelve to load this data.