Introduce TransactionOverview for loading per-quarter transaction totals
We're about to extend ExportActivityToCsv so that it includes the
previous twelve quarters of transaction totals for the activity. The
existing method `Activity#actual_total_for_report_financial_quarter`
method doesn't quite do what we want here -- it takes a Report object,
because like the related `forecasted_total_for_report_financial_quarter`
and `variance_for_report_financial_quarter` methods, it needs to take
per-report versioning into account.

The current logic in ExportActivityToCsv for getting the previous
quarter's transaction total is to:

- use `Report#previous` to find the Report for the quarter before the
  current one
- pass this Report to
  `Activity#actual_total_for_report_financial_quarter` to get the
  transaction total for that quarter

There are some issues with this in terms of consistency. First, it
assumes there is one Report per financial quarter, with no gaps, which
is not true -- some reports do not have a financial quarter, we're
allowed to have multiple reports in the same quarter, and so on.
Although Report *is* ordered according to `Report.in_historical_order`,
that does not mean we have one report per quarter.

Second, `Activity#actual_total_for_report_financial_quarter` looks for
transactions that belong to the given Report *and* are in that report's
date range. It is possible for a Report to include a Transaction that is
from a previous quarter, because that Transaction was not reported at
the time it was made and is being accounted in a later Report.

Therefore, say we have two reports, one in Q2 which is the one
ExportActivityToCsv is based on, and one in Q1 which is the "previous"
Report we use to get the previous quarter's transaction total. If the Q2
report contains a Q1 transaction for that activity, the current
implementation will *not* count it, which seems wrong to me.

Instead, the model we're aiming for is that using Reports to pin
versions of data means you can use a Report to see a snapshot of
whatever state existed when that report was filed. If the Q2 report
contains a Q1 transaction, then the "previous quarter transaction total"
for the Q2 report should include that transaction. If it only includes
transactions from the Q1 report then we're showing a previous report's
snapshot and not including everything included in the Q2 report, which
is confusing.

For all the above reasons, I am introducing a service named
TransactionOverview whose interface mirrors PlannedDisbursementOverview
-- it lets you get a view of per-quarter transaction data, as of a
particular Report. And it does that by including transactions that
appear in the given Report, *or any previous Report* according to
`Report.historically_up_to(report)`.

So, if you want the Q1 transaction total in the Q2 report CSV, this will
include any Q1 transactions that were added in the Q2 report, not just
the Q1 report. But, if the later Q3 report adds more Q1 transactions,
those will not be included in the Q2 report CSV.

This class also includes the ability to do bulk vs single loads of
transaction totals, because we're going to want to load twelve previous
quarters in ExportActivityToCsv. As with PlannedDisbursementOverview,
you can call `value_for_report_quarter` to get the value for the current
report's quarter, or `all_quarters.value_for(quarter)` to do a bulk load
and then select a single quarter's value from there.