Refactor the ExportActivityToCsv tests
In trying to add the previous twelve quarters' transaction totals to the
output of this class, I started to find these tests really hard to work
with. The long lists of "0.00" values don't expose any of the structure
of what this class generates, and making these lists longer results in
useless test failure messages because RSpec starts omitting parts of
them, as in:

    expected: ["apples", "oranges", "bana...ons", "pears"]
         got: ["apples", "oranges", "bana...ons", "pears"]

You can't see the difference if it's somewhere in the middle of the
arrays, especially when what you actually get is:

    expected: ["0.00", "0.00", "0....00", "0.00", "0.00"]
         got: ["0.00", "0.00", "0....00", "0.00", "0.00"]

I've tried to restructure these to better differentiate between
logically distinct parts of the output, and to move the header and value
tests together so it's clearer how they related to each other.