Only create Level D activities with an editable report
Delivery partners can only create Level D activities when there is a
relevant editable report.

Since calling the `policy` view helper with a model instance infers the
policy class from the model (in this case `ActivityPolicy`) with no way
of overriding it, I'm calling `ThirdPartyProjectPolicy.new()`
explicitly.

I've also added a guard clause (with a FIXME comment) in the
`editable_report?` method, so handle cases where this policy is called
with a symbol.

This only ever happens in a helper related to the `level` form step,
which is being removed shortly. After this happens, I want to remove
that piece of code, so calling the policy with a symbol throws an error.