Data migration to remove all legacy data
This data migration will remove all data for known  Delivery Partner
organisations.

The goal of this work is to remove all the data we have imported via the
legacy IATI data for:

- Royal Academy of Engineering
- British Academy

We are doing this because BEIS would like to change the RODA identifier
for all the activities which work 'top down' i.e. any changes to the ID
at the top level need to propagate down so effectively we would need to
change all the IDs. We decided that as the data is to be replaced
entirely (see below) the simplest approach is to remove the data in its
entirety.

It is safe to carry out this action as neither organisation has used RODA
to report as yet and the new data migration process will
replace all of the data with a cleaned and verified data set prepared
for this purpose.

This code has been run against a copy of production data and had the
results verified, we decided, on balance that the change here is at such
a high level that any automated test would not increase our confidence
in this code doing the right thing beyond that of running it locally.

The expectation, as with all data migrations, is that this will be
deleted once it has been deployed and run.