Use TransactionOverview in Activity#actual_total_for_report_financial_quarter
This uses the service we added in the previous commit to change the
implementation of `Activity#actual_total_for_report_financial_quarter`
so that it makes sense with the way financial data is pinned to reports.

This also makes it symmetric with
`Activity#forecasted_total_for_report_financial_quarter` in terms of how
it works, although it's not memoised -- memoising it makes some tests
fail. Since these methods take an argument, it is probably incorrect to
memoise them without indexing on that argument, but I'm not changing
that right now.