Use fund and organisation short names in filename
Currently we use the full title of the source fund, which adds too much noise to the downloaded file’s name.

The requirements are to use the fund short name (“GCRF” / “NF”), as well as the recently added attribute,
`beis_organisation_reference`, which is the organisation's short name.

Also, the financial quarter and year are moved to the start of the filename.