Include the current quarter in actual/forecast groups
Each row of the report CSV is currently ordered so that it lists the
following groups of fields for each activity:

1. All the activity's metadata fields, including the forecast, actual
   and variance for the current financial quarter, and comment

2. Actual spend for the twelve financial quarters before the report's
   own financial quarter

3. Forecasted spend for the twenty financial quarters after the report's
   own quarter

We instead want to organise things so that the current quarter's
forecast and actual spend are part of the series represented by groups 2
and 3, rather than separated off in group 1. We therefore reorganise the
fields as follows:

1. All the activity's metadata fields

2. Actual spend for the twelve financial quarters up to the report's own
   financial quarter

3. Forecasted spend for the next twenty quarters beginning with the
   report's own financial quarter

4. The variance for the report's own financial quarter; the difference
   between the first-listed forecase and the last-listed actual

5. Comments and link to the activity's page

This changes how the forecast and actual values for the current quarter
are fetched. Before they we're part of the "previous 12" or "next 20"
groups but were fetched independently, by calling
`Activity#forecasted_total_for_report_financial_quarter` and
`Activity#actual_total_for_report_financial_quarter`. They're now part
of the broader quarterly series and so are covered by calls to
`PlannedDisbursementOverview#all_quarters` and
`TransactionOverview#all_quarters`.

However, those Activity methods are still called implicitly because the
variance is still fetched by calling
`Activity#variance_for_report_financial_quarter`, which in turn calls
`forecasted_total_for_report_financial_quarter` and
`actual_total_for_report_financial_quarter`.

I believe these results should agree with the values returned by
PlannedDisbursementOverview and TransactionOverview for the given
Report, but it might be nice to remove some of these methods and
calculate the variance off of data we've already fetched for the
quarterly series.