Change submitted email body based on user type
The submitted email will have different bodies based on whether the user is a
delivery partner for the report's organisation, or a service owner.

Rather than have two mailers for each different user type, we have one mailer that
has different content  for each user type.

If, for some reason, the user doesn't belong in either of those  groups, we raise an
error. As we send email asynchronously, this won't effect the experience in the
frontend, but we will be alerted via Rollbar.