Skip zero value transaction rows when importing
Rows where the transaction `Value` is zero are now skipped from importing.

The change in the spec is due to the fact that a row with importer validation errors is not used to attempt to create a transaction, and therefore does not get the model validation performed on it. This means that a row with at least one importer-detected error will not report any model validation errors.