Update sector codes
The goal of the vendor codelist is to enable the team to update the list
with as little effort as possible.

Sector codes have been withdrawn and new ones added and replacing the
two files here shows us what has been changed.

Some of our expectations may have to change to reflect the new codelist.