Show sector code in application
The new sector code code list contains duplicate names, whilst this is
less than ideal for the users, it is something that is outside of the
scope for us to address, the code lists should reflect all of the options
defined by OECD (the body that defines them).

For this reason our simplest response it to add the code to the name to
allow users to make informed decisions about which option to select.

In the past, we must have removed the duplicates as there is a test to
check for duplicates names. This commit removes that spec as we believe
it is better to allow access to all the codes as they are all potential
options for a user, even if the naming is poor.