Document setting AWS and CF credentials for Terraform deployment
This describes how we set these credentials the most recent time we
deployed. We found a few problems with the existing docs:

- You may not have a `~/.aws/credentials` file if you don't use the
  `aws` CLI
- `AWS_ACCESS_KEY` and `AWS_SECRET_ACCESS_KEY` should be in all-caps
- The actual mechanics of how to set these variables in your shell may
  be unclear to inexperienced shell users

I've used a file to prepare the credentials here, rather than recommend
that people paste directly into their shell, which can easily result in
mistakes being made if the credentials contain characters with special
meaning to the shell, such as `;`, `*`, `{`, `$` and so on.