Remove the ability to create funds
In order for a fund activity to function correctly, it needs to be added
alongside so code changes, so it doesn't make sense to add them through
the front end.

Given how the create_activity_level_options helper works, I originally
thought I could acheive this by changing the `create?` method in
`FundPolicy` to `false`. However, the helper checks the result from
`create? || update?`, so it would fall back to `update?` which is
unchanged.

Removing the `update?` check in the helper, caused checks on
`ThirdPartyProjectPolicy` to fail, because this needs to verify that an
active report exists for the associated fund in order to pass, however
this isn't decided until the form step *following* this one(!)

In the end, I decided it would be quicker to remove the `fund` within
the helper method. The helper will go away entirely once we're able to
create activities *only* as a child of an existing activity.

NB: Whilst this does leave the 'level' step intact, with only one option
"programmes (level B)" for a delivery partner, this form step will soon
be removed entirely.