Remove ExtendingOrganisationsController
Changing an activity's `extending_organisation` without affecting the
`organisation` attribute, and those of its children will likely cause
lots of issues.

I believe this was original written to accommodate both programmes being
assigned to the wrong delivery partner at creation, and also to move
activities between delivery partners.

The current thinking is that a programme activity would end on one DP
and then new replacement created at the new DP, in order to maintain the
history correctly.

If we need to offer this functionality in the future, we should consider
what we do with child activities and also the transactions, budgets etc
attached to the activity itself.