Prevent creating third-party projects without a parent
Remove the "third-party projects" radio button from the level form step.
This means the only way to create thse is via the "Add activity" button
on the "child activities" page of an existing activity.

Once this change has been applied at the remaining levels, we will be
able to remove the `level` and `parent` form steps altogether, but I'm
trying to make small steps to avoid breaking lots of tests all at once.