Prevent creating projects without a parent
Similar to the previous commit, only we're removing the ability to
choose "Project (Level C)" from the "level" form step.

This means that for delivery partner users, the level form step will no
longer contain *any* radio buttons, making it impossible to proceed.
This seems fine, since we'll shortly be removing this path altogether.

When rewriting the button code in the `_projects.html.haml` partial to
use `ActivityChildrenController#create`, I noticed that it was using the
current user's organisation, rather than the delivery partner's. This
risked the child activity being assigned to BEIS if a BEIS user somehow
accessed this page.

It makes sense to standardise on `extending_organisation`, as
`organisation` can be inferred using this and `level`, and I'd like to
remove one of them, to prevent duplication.