Use `new_child` to in the CSV importer
This uses our new `Activity.new_child` convenience method to initialize
an activity and set a bunch of defaults based on the Activity's parent.
This eliminates a bunch of cruft from the importer and keeps everything
nice and consistent with how we create activities elsewhere.

I've also added a new `column_to_csv_column_mapping` method to our
`Error` struct to make sure we can give the correct column name when
there is a problem with the parent RODA ID. This was gleaned from the
`Converter::FIELDS` hash before, but now we're no longer setting the
parent in the converter, it's no longer there. We could have just kept
this as `"parent_id"`, but this may have been confusing for the user, as
there's no column with that  name.