Codelist can return a hash of names and codes
We want all the coded values to come from a vendor such as BEIS, IATI,
etc.

To support this we have the Codelist object which creates usable
versions of the code lists in the application.

We often need to go from a named value in user terms to a coded value
and back again.

We often use the locales and the code lists to achieve this, but that
means there are now two places where the code and name are stored, the
locale file and the code list.

By extending Codelist to return simple hashes of the name and codes we
can use them throughout the application where we want to get a readable
name for a specific code or vice versa.

The Codelist class is already written so it only loads the yaml files
once in production to make creating this objects very cheap!