Budget IATI values
Previously we set the `budget_type` and `status` of a Budget when it is
created. This was part of work to change how budgets work in the
application.

The next step is to extend our concept of 'where a budget comes from' -
we have the simple version of this as `funding_type` but this doesn't
really describe what it does, we were forced to use a name other than
`budget_type` as the IATI value was already using that name.

This work uncouples a Budget object from the stored `budget_type` by
loading the types as a constant from a codelist and then returning the
value from a new `iati_type` method, we do the same for `status` with
`iati_status`.

These two methods better describe what is happening by placing them into
a sort of IATI namespace. It also frees the `budget_type` column to
store the new budget types without having to call it 'funding type' -
this should all get tied up in the next couple of commits.