Child activity strategic area is delegated to programme
Project and third-party project activities are delegated up to the
value from the parent level programe.

Fund level activities can't have a GCRF strategic area so we set that to
always be nil.

We will have to ensure that we're eager loading the parent and
grandparent records anytime we're loading in bunch of activities to
avoid n+1 problems.

I had to change the test subject in the CSV importer spec from an
:activity to :programme_activity - I think this is ok. For other levels
the importer shouldn't import any values for strategic area.