Add the ability to set the Report used by ImportPlannedDisbursements
This essentially reverts d37d53fb (Remove the ability to set the Report
used by ImportPlannedDisbursements, 2021-01-21), but required
substantial reworking. Since this functionality was removed, we have
exposed the functionality of this service to end users, adding some
authorization logic and other restrictions.

To make sure this service still works in user-facing contexts, but also
allows developers to back-fill approved reports, the service now accepts
*either* a User or a Report; it is an error to pass both in. If given a
User, the service will automatically import data into the current
editable report and check the user is authorized. If given a Report, it
assumes it's being run by a developer at the command line, and lets us
override which report the data is imported into.